+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 12.21.0
+
+* `[knf]` Added new getter `GetD` which returns valus as duration in seconds
+* `[system/process]` Improved tests
+
 ### 12.20.3
 
 * `go-check` package replaced by [our fork](https://kaos.sh/check)